Fix the gridButton,container height, and create random colors
gridButton = implement a cellSize calculation in js and take the size of userInput number into a loop condition to make it more dynamically created instead in css.
container height issue overlapping the container due to the container not being cleared after userInput a new set of grid size number. Solution: Add a line of code to clear previous gridsize. container.innerHTML =
random color and color hovering - created a randomNumber function with para. of maxnum of the rgb colors and sets it to a getRandomColor() function.